+++ pp/test/test_port_from_csv/test_csv2port.yml
@@ -1,0 +1,14 @@
+E0:
+- ' 10.000'
+- ' 0.000'
+- ' 0'
+- ' 0.500'
+- ' 1'
+- ' 0'
+W0:
+- ' 0.000'
+- ' 0.000'
+- ' 180'
+- ' 0.500'
+- ' 1'
+- ' 0'
